requirements.txt
# Requirements for Deepfake Audio Detection System
# Python 3.8+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
audioread>=3.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0

# Visualization & Logging
matplotlib>=3.7.0
tensorboard>=2.12.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Transformers (optional, for advanced features)
transformers>=4.30.0

# Testing
pytest>=7.3.0

# Utilities
scikit-learn>=1.2.0
